python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "autogen-core==0.4.3",
  "autogen-ext[http]==0.4.3",
  "autogen-agentchat==0.4.3",
  "autogenstudio==0.4.3",
  "openai==1.60.0",
  "tiktoken==0.8.0",
  "ruff>=0.9.2",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "mcp>=1.2.0",
]

[project.optional-dependencies]
jupyter-executor = [
    "ipykernel>=6.29.5",
    "nbclient>=0.10.2",
]

[tool.uv.sources]
autogenstudio = { git = "https://github.com/EItanya/autogen.git", subdirectory = "python/packages/autogen-studio", rev = "c573244206768908f415ea06b415c8af6d1e8d5b" }
autogen-ext = { git = "https://github.com/EItanya/autogen.git", subdirectory = "python/packages/autogen-ext", rev = "c573244206768908f415ea06b415c8af6d1e8d5b" }
kagent = { workspace = true }

[tool.ruff]
line-length = 120
fix = true

target-version = "py310"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "W", "B", "Q", "I", "ASYNC", "T20"]
ignore = ["F401", "E501"]

[tool.hatch.build.targets.wheel]
packages = ["src/kagent"]